Fix large number of file handles open a time when -threads used.
The core issue lies in the fact that, although each file handle is eventually closed, the rate of handle creation by the main thread is significantly faster than the rate of handle closure by the worker threads. This can lead to a temporary buildup of open handles, especially when processing a large number of files. To resolve this, we implement a change to close the handle in the main thread as soon as the hash job is queued in the thread pool, and then have the worker thread reopen the file when its processing begins.